Organized the frontend folder structure
