+++ frontend/public/manifest.json
