+++ frontend/src/App.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function App() {
+  return (
+    <div>
+      <h1>Learn React</h1>
+    </div>
+  );
+}
+
+export default App;
